# services/gateway/internal/domain/paper.go
package domain

type Paper struct {
	Id               string `json:"id"`
	Title            string `json:"title"`
	Abstract         string `json:"abstract"`
	Year             int    `json:"year"`
	Best_oa_location string `json:"best_oa_location"`
	State            string `json:"state"`
}

type ReferencedPaper struct {
	Id string `json:"id"`
}

type RelatedPaper struct {
	Id string `json:"id"`
}